requirements.txt
google-cloud-bigquery==1.28.0